requirements.txt
flask==2.2.3
Flask-SQLAlchemy==3.0.3
psycopg2==2.9.5
bcrypt==4.0.1
PyJWT==2.6.0
datetime==5.1
uuid==1.30